pyproject.toml
[project]
name = "legisguard-ai"
version = "0.1.0"
description = "Local PoC for ingesting Korean safety and health laws into SQLite and Chroma."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "chromadb>=1.0.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "pypdf>=5.2.0",
  "python-multipart>=0.0.9",
  "python-dotenv>=1.0.1",
  "sqlalchemy>=2.0.36",
  "uvicorn>=0.34.0",
]

[project.optional-dependencies]
local-embeddings = [
  "sentence-transformers>=3.4.0",
]
dev = [
  "pytest>=8.3.4",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
